fix/header: fixed links, new component MenuMobile, new hook for handle scroll, new file for links data // fix/hero:fixed background on hover
