+++ src/data/nav-link-data.ts
@@ -1,0 +1,8 @@
+export const navLinks = [
+  { href: '#missao', label: 'Missão' },
+  { href: '#solucoes', label: 'Soluções' },
+  { href: '#diferenciais', label: 'Diferenciais' },
+  { href: '#contato', label: 'Contato' },
+  { href: '/blog', label: 'Blog' },
+]
+export default navLinks
